[Refactoring] : MVVM architecture
기존 회원가입 부분 로직을 ViewModel사용하여 MVVM 아키텍처로 리팩토링
진행
UI 상태 데이터만 ViewModel 로 분리하여 Observer 달아놓았으며 Usecase
단위 도입 및 DI 코드 작성
아직 API 호출 부분은 분리하지 않았음
