+++ app/build.gradle
@@ -41,6 +41,8 @@ android {
 
 dependencies {
     def room_version = "2.5.1"
+    def lifecycle_version = "2.5.1"
+    def arch_version = "2.1.0"
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
@@ -49,6 +51,9 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    implementation 'androidx.activity:activity-ktx:1.2.0'
+    implementation 'androidx.fragment:fragment-ktx:1.3.0'
+
 
 
     // 카드뷰 사용
@@ -82,7 +87,12 @@ dependencies {
     // SimpleRatingBar
     implementation 'com.github.ome450901:SimpleRatingBar:1.5.1'
 
-
     implementation "com.kakao.sdk:v2-user:2.19.0" // 카카오 로그인
+
+    // ViewModel lifecycle
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
+    // LiveData
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
+
 
 }
