--- app/src/main/java/com/example/beering/feature/auth/join/JoinViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.beering.feature.auth.join
-
-import androidx.lifecycle.ViewModel
-
-class JoinViewModel : ViewModel() {
-
-}
