+++ app/src/main/java/com/example/beering/feature/auth/join/domain/NameValidations.kt
@@ -1,0 +1,7 @@
+package com.example.beering.feature.auth.join.domain
+
+data class NameValidations(
+    val characters : Boolean,
+    val length : Boolean,
+    val valid : Boolean
+)
