+++ app/src/main/java/com/example/beering/feature/auth/join/domain/PwValidations.kt
@@ -1,0 +1,9 @@
+package com.example.beering.feature.auth.join.domain
+
+data class PwValidations(
+    val englishChars : Boolean,
+    val specialChars : Boolean,
+    val numbers : Boolean,
+    val length : Boolean,
+    val valid : Boolean
+)
